fix: notification toggle (#1806)
* fix: notification toggle consistency

* fix: import issue
